Can confirm driver but I had a crash so I am going to try and recreate the error.
