+++ app/src/main/res/layout/offeringdriverslist_item.xml
@@ -2,20 +2,9 @@
 <GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="40dp"
+    android:layout_height="60dp"
     android:clickable="false"
     android:focusable="false">
-
-    <RadioButton
-        android:text="Confirm Driver"
-        android:layout_row="0"
-        android:layout_column="3"
-        android:gravity="center"
-        android:lines="1"
-        android:textAlignment="center"
-        android:layout_width="160dp"
-        android:layout_height="40dp"
-        android:id="@+id/confirmDriverRadioButton" />
 
     <comcmput301f16t01.github.carrier.UsernameTextView
         android:text=""
@@ -34,6 +23,13 @@
         android:lines="1"
         android:scrollHorizontally="true"
         android:ellipsize="end"/>
+
+    <Button
+        android:text="Confirm Driver"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/confirmDriverButton"
+        android:clickable="false" />
 
     <ImageButton
         android:layout_column="1"
